serf/requirements.txt
backports.csv==1.0.7
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-allauth==0.39.1
django-bootstrap3==11.0.0
django-constance==1.3.4
django-crispy-forms==1.7.2
django-import-export==0.7.0
django-paintstore==0.2
django-picklefield==0.3.2
django-suit==0.2.26
django-widget-tweaks==1.4.5
et-xmlfile==1.0.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
oauthlib==3.0.2
odfpy==1.4.0
openpyxl==2.6.2
pdfrw==0.4
Pillow==6.1.0
pycparser==2.19
Pyphen==0.9.5
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
tinycss2==1.0.2
urllib3==1.25.3
WeasyPrint==0.42.3
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
